新增多个配置文件，更新文件路径格式，并删除不再需要的城市数据文件和Excel处理工具文件

+++ vue3_projects/utils/copy_utils.py
@@ -12,7 +12,8 @@ def do_copy(file_path, config):
     dst = '{project_path}{file_path}'.format(
         project_path=config.target_project_path, file_path=file_path).replace('{sign}', config.target_package_sign)
     if os.path.isdir(src):
-        rmtree(dst)
+        if os.path.isdir(dst):
+            rmtree(dst)
         mkdir(dst)
         copytree(src, dst,  dirs_exist_ok=True)
     else:
